requirements_fixed.txt
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==3.7.1
async-timeout==4.0.3
attrs==25.3.0
av==14.4.0
backoff==2.2.1
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
brotlipy==0.7.0
build==1.2.2.post1
cachetools==5.5.2
certifi
cffi==1.17.1
chardet
charset-normalizer
chromadb==1.0.15
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
cryptography==45.0.3
dataclasses-json==0.6.7
deepgram-sdk==3.3.2
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
dotenv==0.9.9
durationpy==0.10
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fastapi==0.104.1
filelock
filetype==1.2.0
Flask==3.0.3
flask-cors==5.0.1
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec
gcloud==0.18.3
genesis==0.4.1
gmpy2
google-adk==0.0.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.93.1
google-cloud-bigquery==3.33.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-texttospeech==2.26.0
google-crc32c==1.7.1
google-genai==1.4.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.71.0
grpcio-tools==1.73.1
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.29.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-google-genai==2.0.10
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langgraph==0.4.5
langgraph-checkpoint==2.0.26
langgraph-checkpoint-sqlite==2.0.10
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.69
langsmith==0.3.45
livekit>=1.0.12
livekit-agents==1.0.21
livekit-api==1.0.2
livekit-plugins-cartesia==0.4.10
livekit-plugins-deepgram==1.0.21
livekit-plugins-google==1.0.17
livekit-plugins-noise-cancellation==0.2.3
livekit-plugins-openai==1.0.17
livekit-plugins-silero==1.0.21
livekit-plugins-tavus==1.0.22
livekit-plugins-turn-detector==1.0.21
livekit-protocol==1.0.2
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe
marshmallow==3.26.1
mdurl==0.1.2
mem0ai==0.1.102
mmh3==5.1.0
mpmath
multidict==6.4.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.3.1
onnxruntime==1.19.2
openai==1.75.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.18
ormsgpack==1.9.1
overrides==7.7.0
packaging==24.2
pandas==2.3.0
pika==1.3.2
pillow
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
playwright==1.53.0
pluggy==1.5.0
portalocker==2.10.1
posthog==4.2.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pycosat
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.10.1
pydantic_core==2.33.1
pyee==13.0.0
Pygments==2.19.0
PyJWT==2.10.1
pymongo==4.12.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks
pytest==8.3.5
pytest-asyncio==1.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.14.2
rabbitmq-admin==0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.26.0
rsa==4.9.1
ruamel-yaml-conda
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.0.0
shapely==2.0.7
shellingham==1.5.4
simple-websocket==1.1.0
six
sniffio==1.3.1
sounddevice==0.5.1
SQLAlchemy==2.0.41
sqlite-vec==0.1.6
starlette==0.27.0
sympy
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.67.1
transformers==4.49.0
typer==0.16.0
types-protobuf==4.25.0.20240417
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.23.2
uvloop==0.21.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
